wines-function/requirements.txt
pyrebase4
pandas